refactor: extract community functions into dedicated module
