+++ utils/nostr/nostr-helper-functions.ts
@@ -10,8 +10,6 @@ import {
 import { v4 as uuidv4 } from "uuid";
 import CryptoJS from "crypto-js";
 import {
-  Community,
-  CommunityRelays,
   NostrEvent,
   ProductFormValues,
   SavedAddress,
@@ -872,188 +870,6 @@ export async function publishSpendingHistoryEvent(
   }
 }
 
-export async function createOrUpdateCommunity(
-  signer: NostrSigner,
-  nostr: NostrManager,
-  details: {
-    d: string; // The unique identifier, should be constant for a community
-    name: string;
-    description: string;
-    image: string;
-    moderators: string[];
-    relays?: CommunityRelays; // optional relay declarations
-  }
-) {
-  const tags: string[][] = [
-    ["d", details.d],
-    ["name", details.name],
-    ["description", details.description],
-    ["image", details.image],
-    ["t", "shopstr"],
-  ];
-
-  // moderators as p tags with role marker
-  for (const mod_pk of details.moderators) {
-    tags.push(["p", mod_pk, "", "moderator"]);
-  }
-
-  // include relay tags if provided: ["relay", url, type]
-  if (details.relays) {
-    const {
-      approvals = [],
-      requests = [],
-      metadata = [],
-      all = [],
-    } = details.relays;
-    for (const r of approvals) tags.push(["relay", r, "approvals"]);
-    for (const r of requests) tags.push(["relay", r, "requests"]);
-    for (const r of metadata) tags.push(["relay", r, "metadata"]);
-    for (const r of all) tags.push(["relay", r]);
-  }
-
-  const eventTemplate: EventTemplate = {
-    kind: 34550,
-    created_at: Math.floor(Date.now() / 1000),
-    tags,
-    content: "",
-  };
-
-  const signedEvent = await finalizeAndSendNostrEvent(
-    signer,
-    nostr,
-    eventTemplate
-  );
-  // Cache community event to database
-  if (signedEvent) {
-    await cacheEventToDatabase(signedEvent).catch((error) =>
-      console.error("Failed to cache community event to database:", error)
-    );
-  }
-  return signedEvent;
-}
-
-export async function createCommunityPost(
-  signer: NostrSigner,
-  nostr: NostrManager,
-  community: Community,
-  content: string,
-  options?: {
-    parentEvent?: NostrEvent; // reply
-    crosspostCommunities?: Community[]; // cross-post to other communities (NIP-18)
-    externalId?: string; // NIP-73 external content id
-    contentKind?: string; // optional k/i usage
-  }
-) {
-  const communityAddress = `${community.kind}:${community.pubkey}:${community.d}`;
-  const tags: string[][] = [];
-
-  // Always include uppercase A/P tags pointing to the community address/pubkey
-  tags.push(["A", communityAddress]);
-  tags.push(["P", community.pubkey]);
-  tags.push(["K", String(community.kind)]);
-
-  if (options?.parentEvent) {
-    // reply: reference parent event via e and p tags and include parent kind
-    tags.push(["a", communityAddress]);
-    tags.push(["e", options.parentEvent.id, ""]);
-    tags.push(["p", options.parentEvent.pubkey, ""]);
-    tags.push(["k", String(options.parentEvent.kind)]);
-  } else {
-    // top-level announcement: include lowercase a/p/k pointing to the community address/kind
-    tags.push(["a", communityAddress]);
-    tags.push(["p", community.pubkey]);
-    tags.push(["k", String(community.kind)]);
-  }
-
-  // cross-posting: include additional lowercase a tags pointing to other communities' addresses
-  if (options?.crosspostCommunities) {
-    for (const c of options.crosspostCommunities) {
-      const addr = `${c.kind}:${c.pubkey}:${c.d}`;
-      tags.push(["a", addr]);
-    }
-  }
-
-  // NIP-73 external ID support (i tag)
-  if (options?.externalId) {
-    tags.push(["i", options.externalId]);
-    if (options?.contentKind) tags.push(["k", options.contentKind]);
-  }
-
-  const eventTemplate: EventTemplate = {
-    kind: 1111,
-    created_at: Math.floor(Date.now() / 1000),
-    tags,
-    content,
-  };
-
-  // returns signed event (so caller can know id)
-  const signedEvent = await finalizeAndSendNostrEvent(
-    signer,
-    nostr,
-    eventTemplate
-  );
-  // Cache community post event to database
-  if (signedEvent) {
-    await cacheEventToDatabase(signedEvent).catch((error) =>
-      console.error("Failed to cache community post event to database:", error)
-    );
-  }
-  return signedEvent;
-}
-
-export async function approveCommunityPost(
-  signer: NostrSigner,
-  nostr: NostrManager,
-  postToApprove: NostrEvent,
-  community: Community
-) {
-  const communityAddress = `${community.kind}:${community.pubkey}:${community.d}`;
-  const tags: string[][] = [
-    ["a", communityAddress],
-    ["e", postToApprove.id],
-    ["p", postToApprove.pubkey],
-    ["k", String(postToApprove.kind)],
-  ];
-  const eventTemplate: EventTemplate = {
-    kind: 4550,
-    created_at: Math.floor(Date.now() / 1000),
-    tags,
-    content: JSON.stringify(postToApprove),
-  };
-
-  // returns signed approval event (so caller can persist approval id)
-  const signedEvent = await finalizeAndSendNostrEvent(
-    signer,
-    nostr,
-    eventTemplate
-  );
-  // Cache community approval event to database
-  if (signedEvent) {
-    await cacheEventToDatabase(signedEvent).catch((error) =>
-      console.error(
-        "Failed to cache community approval event to database:",
-        error
-      )
-    );
-  }
-  return signedEvent;
-}
-
-// Moderator retract of approval -> publish deletion event (NIP-09, kind 5)
-export async function retractApproval(
-  signer: NostrSigner,
-  nostr: NostrManager,
-  approvalEventId: string,
-  reason?: string
-) {
-  const eventTemplate: EventTemplate = {
-    kind: 5,
-    created_at: Math.floor(Date.now() / 1000),
-    tags: [["e", approvalEventId]],
-    content: reason || `Retract approval ${approvalEventId}`,
-  };
-  return await finalizeAndSendNostrEvent(signer, nostr, eventTemplate);
-}
 
 type FinalizeAndSendOptions = {
   waitForRelayPublish?: boolean;
